Fixed alomst everything, please sort out the JOption panes if you get time they are a little bit of a mess

+++ Eclipse Code Only/src/region/Region.java
@@ -2,6 +2,7 @@ package region;
 public interface Region   
 {
 		
-	public float getVat(Float subTotal);
+
+	public double getVat(double subTotal);
 
 }
